# frontend/src/lib/openapi.json
{
  "components": {
    "schemas": {
...
      "AnswerOutcome": {
        "enum": [
          "answered",
          "abstained",
          "partial"
        ],
        "title": "AnswerOutcome",
        "type": "string"
      },
...
      "CitationOut": {
        "properties": {
          "amends": {
            "items": {
              "$ref": "#/components/schemas/DocRef"
            },
            "title": "Amends",
            "type": "array"
          },
          "branch_specific": {
            "default": false,
            "title": "Branch Specific",
            "type": "boolean"
          },
          "chunk_id": {
            "format": "uuid",
            "title": "Chunk Id",
            "type": "string"
          },
          "doc_code": {
            "title": "Doc Code",
            "type": "string"
          },
          "doc_type": {
            "title": "Doc Type",
            "type": "string"
          },
          "document_id": {
            "format": "uuid",
            "title": "Document Id",
            "type": "string"
          },
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "heading": {
            "title": "Heading",
            "type": "string"
          },
          "marker": {
            "title": "Marker",
            "type": "string"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "section_path": {
            "title": "Section Path",
            "type": "string"
          },
          "snippet": {
            "title": "Snippet",
            "type": "string"
          },
          "supersedes": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DocRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "support_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Support Score"
          },
          "supported": {
            "title": "Supported",
            "type": "boolean"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "version_id": {
            "format": "uuid",
            "title": "Version Id",
            "type": "string"
          }
        },
        "required": [
          "marker",
          "chunk_id",
          "version_id",
          "document_id",
          "doc_code",
          "title",
          "doc_type",
          "version",
          "section_path",
          "heading",
          "page",
          "effective_from",
          "supersedes",
          "amends",
          "snippet",
          "supported",
          "support_score",
          "branch_specific"
        ],
        "title": "CitationOut",
        "type": "object"
      },
...
      "ConflictOut": {
        "properties": {
          "a": {
            "$ref": "#/components/schemas/ConflictSide"
          },
          "b": {
            "$ref": "#/components/schemas/ConflictSide"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "detected_by": {
            "default": "query",
            "title": "Detected By",
            "type": "string"
          },
          "flagged_to_owner": {
            "default": false,
            "title": "Flagged To Owner",
            "type": "boolean"
          },
          "id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "newer": {
            "anyOf": [
              {
                "enum": [
                  "a",
                  "b"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Newer"
          },
          "status": {
            "default": "open",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "description",
          "a",
          "b",
          "newer",
          "status",
          "detected_by",
          "flagged_to_owner"
        ],
        "title": "ConflictOut",
        "type": "object"
      },
...
      "ConflictSide": {
        "properties": {
          "chunk_id": {
            "format": "uuid",
            "title": "Chunk Id",
            "type": "string"
          },
          "doc_code": {
            "title": "Doc Code",
            "type": "string"
          },
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "marker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Marker"
          },
          "section_path": {
            "title": "Section Path",
            "type": "string"
          },
          "snippet": {
            "title": "Snippet",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "marker",
          "chunk_id",
          "doc_code",
          "title",
          "version",
          "section_path",
          "effective_from",
          "snippet"
        ],
        "title": "ConflictSide",
        "type": "object"
      },
...
      "ContactOut": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "pager": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pager"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          },
          "phone_ext": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone Ext"
          },
          "role_label": {
            "title": "Role Label",
            "type": "string"
          }
        },
        "required": [
          "id",
          "role_label",
          "phone_ext",
          "phone",
          "pager",
          "notes"
        ],
        "title": "ContactOut",
        "type": "object"
      },
...
      "DocRef": {
        "properties": {
          "doc_code": {
            "title": "Doc Code",
            "type": "string"
          },
          "section_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Section Path"
          }
        },
        "required": [
          "doc_code",
          "section_path"
        ],
        "title": "DocRef",
        "type": "object"
      },
...
      "EscalationOut": {
        "properties": {
          "clarifying_question": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clarifying Question"
          },
          "contacts": {
            "items": {
              "$ref": "#/components/schemas/ContactOut"
            },
            "title": "Contacts",
            "type": "array"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "reason": {
            "enum": [
              "not_found",
              "high_risk",
              "out_of_scope",
              "clarify",
              "unavailable"
            ],
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason",
          "message",
          "contacts",
          "clarifying_question"
        ],
        "title": "EscalationOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HistoryItem": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnswerOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "query_id": {
            "format": "uuid",
            "title": "Query Id",
            "type": "string"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "route": {
            "$ref": "#/components/schemas/QueryRoute"
          }
        },
        "required": [
          "query_id",
          "question",
          "route",
          "outcome",
          "created_at"
        ],
        "title": "HistoryItem",
        "type": "object"
      },
...
      "QueryRequest": {
        "properties": {
          "branch_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Branch Id"
          },
          "question": {
            "maxLength": 1000,
            "minLength": 2,
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "QueryRequest",
        "type": "object"
      },
      "QueryResponse": {
        "properties": {
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "citations": {
            "items": {
              "$ref": "#/components/schemas/CitationOut"
            },
            "title": "Citations",
            "type": "array"
          },
          "conflicts": {
            "items": {
              "$ref": "#/components/schemas/ConflictOut"
            },
            "title": "Conflicts",
            "type": "array"
          },
          "disclaimer": {
            "default": "Supports, does not replace, clinical judgement.",
            "title": "Disclaimer",
            "type": "string"
          },
          "escalation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EscalationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "generation_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generation Mode"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "outcome": {
            "$ref": "#/components/schemas/AnswerOutcome"
          },
          "pii_redacted": {
            "default": false,
            "title": "Pii Redacted",
            "type": "boolean"
          },
          "query_id": {
            "format": "uuid",
            "title": "Query Id",
            "type": "string"
          },
          "quick_values": {
            "items": {
              "$ref": "#/components/schemas/QuickValue"
            },
            "title": "Quick Values",
            "type": "array"
          },
          "redacted_question": {
            "title": "Redacted Question",
            "type": "string"
          },
          "route": {
            "$ref": "#/components/schemas/QueryRoute"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/SourceCard"
            },
            "title": "Sources",
            "type": "array"
          },
          "timings_ms": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Timings Ms",
            "type": "object"
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "query_id",
          "route",
          "outcome",
          "answer",
          "citations",
          "quick_values",
          "conflicts",
          "escalation",
          "sources",
          "generation_mode",
          "verification",
          "redacted_question",
          "pii_redacted",
          "latency_ms",
          "timings_ms",
          "disclaimer"
        ],
        "title": "QueryResponse",
        "type": "object"
      },
      "QueryRoute": {
        "enum": [
          "answer",
          "clarify",
          "out_of_scope",
          "high_risk"
        ],
        "title": "QueryRoute",
        "type": "string"
      },
      "QuickValue": {
        "properties": {
          "label": {
            "title": "Label",
            "type": "string"
          },
          "source": {
            "title": "Source",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "label",
          "value",
          "source"
        ],
        "title": "QuickValue",
        "type": "object"
      },
...
      "SourceCard": {
        "properties": {
          "amends": {
            "items": {
              "$ref": "#/components/schemas/DocRef"
            },
            "title": "Amends",
            "type": "array"
          },
          "branch_specific": {
            "default": false,
            "title": "Branch Specific",
            "type": "boolean"
          },
          "chunk_id": {
            "format": "uuid",
            "title": "Chunk Id",
            "type": "string"
          },
          "doc_code": {
            "title": "Doc Code",
            "type": "string"
          },
          "doc_type": {
            "title": "Doc Type",
            "type": "string"
          },
          "effective_from": {
            "format": "date",
            "title": "Effective From",
            "type": "string"
          },
          "heading": {
            "title": "Heading",
            "type": "string"
          },
          "marker": {
            "title": "Marker",
            "type": "string"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "relevance": {
            "title": "Relevance",
            "type": "number"
          },
          "section_path": {
            "title": "Section Path",
            "type": "string"
          },
          "snippet": {
            "title": "Snippet",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "version_id": {
            "format": "uuid",
            "title": "Version Id",
            "type": "string"
          }
        },
        "required": [
          "marker",
          "chunk_id",
          "version_id",
          "doc_code",
          "title",
          "doc_type",
          "version",
          "section_path",
          "heading",
          "page",
          "effective_from",
          "snippet",
          "relevance",
          "amends",
          "branch_specific"
        ],
        "title": "SourceCard",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VerificationOut": {
        "properties": {
          "claims": {
            "title": "Claims",
            "type": "integer"
          },
          "judge": {
            "title": "Judge",
            "type": "string"
          },
          "supported": {
            "title": "Supported",
            "type": "integer"
          }
        },
        "required": [
          "claims",
          "supported",
          "judge"
        ],
        "title": "VerificationOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Source-backed answers from approved institutional documents. ProtoCite retrieves and displays approved documents; it does not provide diagnoses or patient-specific treatment or dosing recommendations.",
    "title": "ProtoCite API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/query": {
      "post": {
        "operationId": "ask_api_v1_query_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask",
        "tags": [
          "query"
        ]
      }
    },
    "/api/v1/query/history": {
      "get": {
        "operationId": "history_api_v1_query_history_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HistoryItem"
                  },
                  "title": "Response History Api V1 Query History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "History",
        "tags": [
          "query"
        ]
      }
    },
    "/api/v1/query/stream": {
      "get": {
        "description": "SSE over GET (spec compatibility). Access logs record the path only, never the query string.",
        "operationId": "ask_stream_get_api_v1_query_stream_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 1000,
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "branch_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Branch Id"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask Stream Get",
        "tags": [
          "query"
        ]
      },
      "post": {
        "description": "Preferred streaming form: the question travels in the body, never in a URL.",
        "operationId": "ask_stream_post_api_v1_query_stream_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ask Stream Post",
        "tags": [
          "query"
        ]
      }
    },
...
  }
}